delete. コンポーネント内のファイル add. pageファイルにデータ取得の記述
