--- src/components/livePage.tsx
@@ -1,5 +1,0 @@
-const LivePage = () => {
-  return <div>this is LivePage</div>;
-};
-
-export default LivePage;
