--- src/components/stagePage.tsx
@@ -1,5 +1,0 @@
-const StagePage = () => {
-  return <div>this is StagePage</div>;
-};
-
-export default StagePage;
